requirements.txt
streamlit
google-generativeai
requests
dotenv